+++ include/fuse_common.h
@@ -206,7 +206,6 @@ struct fuse_conn_info {
 };
 
 struct fuse_session;
-struct fuse_chan;
 struct fuse_pollhandle;
 
 /**